requirements.txt
metaphone
pandas
pytest
fuzzymatcher
fuzzywuzzy
dvc
boto3
datatest
allure-pytest
rapidfuzz
openpyxl
#npm install allure-commandline --save-dev

